Update spec and code to use quoted vs unquoted
This allows us to dispense with server side guessing
of whether a value needs to be quoted. This puts the
onus on the client, which anyway knows better what type
of value this is.

We can have an explicit type in the Python wrapper to
take care of this.

Remove deprecated isNumeric function