Re-org of query API to support asnc queries (#76)
The Query APIs are all now async, and require one API call to create the query and another to get the query results.

This change modifies the `bulk-data`, `boolean-query` and `sql-query` endpoints. Instead of returning the query result, they now return a query ID. To get the query results, you must pass the ID to new endpoint, `query`. Note that the pagination inputs `offset` and `limit` are now parameters to `query`.

The query endpoint returns a `next_url ` as a convenience to return the next page of data.

The `query` will return an empty result if a query is not yet complete in BigQuery, so users will need to poll this endpoint to get their query results. Once the query is complete, all pages of data are present.

This is a breaking change that will require a new version of https://github.com/CancerDataAggregator/cda-python to work once it's deployed to the dev instance.

Fixes #14, Fixes #58